% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionality.R
\name{convertSigma}
\alias{convertSigma}
\title{Convert between CLR and ALR covariance matrices}
\usage{
convertSigma(S, direction = c("alr2clr", "clr2alr"))
}
\arguments{
\item{S}{Covariance matrix to be converted}

\item{direction}{Which direction to convert between alr and clr.}
}
\value{
A covariance matrix on the requested scale.
}
\description{
Convert between CLR and ALR covariance matrices
}
\examples{
convertSigma(diag(3), "alr2clr")
}
