% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlecell-data.R
\docType{data}
\name{singlecell}
\alias{singlecell}
\title{Single cell sequencing data from mouse embryonic stem cells in G1 phase}
\format{
## `singlecell`
A matrix with 96 rows and 10 columns.
}
\source{
<https://www.ebi.ac.uk/biostudies/arrayexpress/studies/E-MTAB-2805>
}
\usage{
data(singlecell)
}
\description{
A subset of single cell data from Buettner et al. 2015.  Contains single
cell measurements from 96 mouse embryonic stem cells all in G1 phase.
}
\examples{
data(singlecell)

}
\keyword{datasets}
