% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{wrapMLE}
\alias{wrapMLE}
\title{Wrapper for mleLR()}
\usage{
wrapMLE(x)
}
\arguments{
\item{x}{Input argument for the \code{MLE Function}}
}
\description{
Executes a function call to \code{mlePath()}. Helps with the \code{mclapply()}
within \code{mlePath()}
}
