% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionality.R
\name{logitNormalVariation}
\alias{logitNormalVariation}
\title{Logit Normal Variation}
\usage{
logitNormalVariation(
  mu,
  Sigma,
  type = c("standard", "phi", "phis", "rho"),
  order = c("second", "first")
)
}
\arguments{
\item{mu}{The mle estimate of the mu matrix}

\item{Sigma}{The mle estimate of the Sigma matrix}

\item{type}{Type of variation metric to be calculated: \code{standard}, \code{phi},
\code{phis} (a symmetrical version of \code{phi}), or \code{rho}}

\item{order}{The order of the Taylor-series approximation to be used in the
estimation}
}
\value{
An estimation of the variation matrix, \code{V}.
}
\description{
Estimates the variation matrix of count-compositional data
based on a the same approximation used in logitNormalVariation()
only for this function it uses empirical estimates of mu and Sigma.
Also performs zero-imputation using \code{cmultRepl()} from the
\code{zCompositions} package.
}
\examples{
data(singlecell)
mle <- mleLR(singlecell)
mu.hat <- mle$mu
Sigma.hat <- mle$Sigma

logitNormalVariation(mu.hat, Sigma.hat)
logitNormalVariation(mu.hat, Sigma.hat, type="phi")
logitNormalVariation(mu.hat, Sigma.hat, type="rho")

}
