% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionality.R
\name{pluginVariation}
\alias{pluginVariation}
\title{Plugin Variation}
\usage{
pluginVariation(
  counts,
  type = c("standard", "phi", "phis", "rho"),
  order = c("second", "first"),
  impute.zeros = TRUE,
  ...
)
}
\arguments{
\item{counts}{Matrix of counts; samples are rows and features are columns.}

\item{type}{Type of variation metric to be calculated: \code{standard}, \code{phi},
\code{phis} (a symmetrical version of \code{phi}), or \code{rho}}

\item{order}{The order of the Taylor-series approximation to be used in the
estimation}

\item{impute.zeros}{If TRUE, then \code{cmultRepl()} from the \code{zCompositions} package is used to impute zero values in the counts matrix.}

\item{...}{Optional arguments passed to zero-imputation function \code{cmultRepl()}}
}
\value{
An estimation of the variation matrix, \code{V}.
}
\description{
Estimates the variation matrix of count-compositional data
based on a multinomial logit-Normal distribution. Estimate is performed using
only the parameters of the distribution.
}
\examples{
data(singlecell)

pluginVariation(singlecell)
pluginVariation(singlecell, type="phi")
pluginVariation(singlecell, type="rho")

}
